third_party/blink/web_tests/platform/generic/fast/parser/append-child-followed-by-document-write-expected.txt

ALERT: document.documentElement.outerHTML: <a></a>
document.childNodes.length: 1
b element: null
